+++ Tests/ProjectDescriptionTests/SchemeTests.swift
@@ -13,18 +13,18 @@ final class SchemeTests: XCTestCase {
                                                       preActions: buildAction,
                                                       postActions: buildAction),
                              testAction: TestAction(targets: ["target"],
-                                                    arguments: Arguments(environment: ["a": "b"],
-                                                                         launch: ["a": true]),
+                                                    arguments: Arguments(environment: ["test": "b"],
+                                                                         launch: ["test": true]),
                                                     config: .debug,
                                                     coverage: true,
                                                     preActions: testAction,
                                                     postActions: testAction),
                              runAction: RunAction(config: .debug,
                                                   executable: "executable",
-                                                  arguments: Arguments(environment: ["a": "b"],
-                                                                       launch: ["a": true])))
+                                                  arguments: Arguments(environment: ["run": "b"],
+                                                                       launch: ["run": true])))
 
-        let expected = "{\"build_action\": {\"targets\": [\"target\"], \"pre_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo build_action\", \"target\": \"target\"}], \"post_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo build_action\", \"target\": \"target\"}]}, \"name\": \"scheme\", \"run_action\": {\"arguments\": {\"environment\": {\"a\": \"b\"}, \"launch\": {\"a\": true}}, \"config\": \"debug\", \"executable\": \"executable\"}, \"shared\": true, \"test_action\": {\"arguments\": {\"environment\": {\"a\": \"b\"}, \"launch\": {\"a\": true}}, \"config\": \"debug\", \"coverage\": true, \"targets\": [\"target\"],  \"pre_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo test_action\", \"target\": \"target\"}], \"post_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo test_action\", \"target\": \"target\"}]}}"
+        let expected = "{\"build_action\": {\"targets\": [\"target\"], \"pre_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo build_action\", \"target\": \"target\"}], \"post_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo build_action\", \"target\": \"target\"}]}, \"name\": \"scheme\", \"run_action\": {\"arguments\": {\"environment\": {\"run\": \"b\"}, \"launch\": {\"run\": true}}, \"config\": \"debug\", \"executable\": \"executable\"}, \"shared\": true, \"test_action\": {\"arguments\": {\"environment\": {\"test\": \"b\"}, \"launch\": {\"test\": true}}, \"config\": \"debug\", \"coverage\": true, \"targets\": [\"target\"],  \"pre_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo test_action\", \"target\": \"target\"}], \"post_actions\": [{\"title\": \"Run Script\", \"script_text\": \"echo test_action\", \"target\": \"target\"}]}}"
         assertCodableEqualToJson(subject, expected)
     }
 }